Refactor handler registration
